# dist/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "PayGate Aggregator API",
    "version": "1.0.0",
    "description": "Unified Payment Gateway Aggregation API. Integrate Stripe, PayPal, and Mercado Pago through a single normalized interface.",
    "contact": {
      "name": "API Support",
      "email": "support@paygate.io"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development"
    },
    {
      "url": "https://api.paygate.io",
      "description": "Production"
    },
    {
      "url": "/",
      "description": "Same origin (use with Try it out)"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "BearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key as Bearer token"
      }
    },
    "schemas": {
      "CreatePaymentRequest": {
        "type": "object",
        "required": [
          "provider",
          "amount",
          "currency",
          "description",
          "customer"
        ],
        "description": "Provide exactly one of paymentMethod (new token/pm/tok) or savedCardId (from GET /api/v1/cards).",
        "properties": {
          "provider": {
            "type": "string",
            "enum": [
              "stripe",
              "paypal",
              "mercadopago"
            ]
          },
          "amount": {
            "type": "integer",
            "description": "Amount in smallest currency unit (cents/centavos)",
            "example": 1500
          },
          "currency": {
            "type": "string",
            "example": "ARS",
            "description": "ISO 4217 currency code"
          },
          "description": {
            "type": "string",
            "example": "Order #8821"
          },
          "customer": {
            "type": "object",
            "required": [
              "email",
              "name"
            ],
            "properties": {
              "email": {
                "type": "string",
                "format": "email"
              },
              "name": {
                "type": "string"
              }
            }
          },
          "paymentMethod": {
            "type": "object",
            "required": [
              "type",
              "token"
            ],
            "properties": {
              "type": {
                "type": "string",
                "enum": [
                  "card",
                  "bank_transfer",
                  "wallet"
                ]
              },
              "token": {
                "type": "string",
                "description": "pm_..., tok_visa, etc."
              }
            }
          },
          "savedCardId": {
            "type": "string",
            "description": "Id from POST /api/v1/cards — pay with that saved card (Stripe only; customer.email must match)."
          },
          "metadata": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "PaymentResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "pay_01J3X8K9MABCDEF12345"
          },
          "provider": {
            "type": "string",
            "enum": [
              "stripe",
              "paypal",
              "mercadopago"
            ]
          },
          "status": {
            "type": "string",
            "enum": [
              "pending",
              "processing",
              "succeeded",
              "failed",
              "cancelled"
            ]
          },
          "amount": {
            "type": "integer"
          },
          "currency": {
            "type": "string"
          },
          "providerTransactionId": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "metadata": {
            "type": "object"
          }
        }
      },
      "PaymentListItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "provider": {
            "type": "string",
            "enum": [
              "mercadopago"
            ]
          },
          "mpPaymentId": {
            "type": "string"
          },
          "externalReference": {
            "type": "string",
            "nullable": true
          },
          "preferenceId": {
            "type": "string",
            "nullable": true
          },
          "merchantOrderId": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "string"
          },
          "statusDetail": {
            "type": "string",
            "nullable": true
          },
          "amount": {
            "type": "string",
            "nullable": true,
            "description": "Transaction amount as decimal string (Mercado Pago)"
          },
          "currency": {
            "type": "string",
            "nullable": true
          },
          "payerEmail": {
            "type": "string",
            "nullable": true
          },
          "paymentMethodId": {
            "type": "string",
            "nullable": true
          },
          "paymentTypeId": {
            "type": "string",
            "nullable": true
          },
          "dateApproved": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "PaymentListResponse": {
        "type": "object",
        "properties": {
          "payments": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentListItem"
            }
          },
          "total": {
            "type": "integer"
          },
          "limit": {
            "type": "integer"
          },
          "offset": {
            "type": "integer"
          }
        }
      },
...
    },
    "responses": {
      "ValidationError": {
        "description": "Validation error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "details": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        }
      },
      "Unauthorized": {
        "description": "Unauthorized",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
...
      "ProviderError": {
        "description": "Payment provider error",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "properties": {
                "error": {
                  "type": "string"
                },
                "code": {
                  "type": "string"
                },
                "provider": {
                  "type": "string",
                  "example": "stripe"
                }
              }
            }
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/v1/payments": {
      "post": {
        "tags": [
          "Payments"
        ],
        "summary": "Create a payment",
        "description": "Charges the selected provider. When `API_KEY_SECRET` is set on the server, send it as a Bearer token.\n\n**Stripe test mode:** use `paymentMethod.token` = `tok_visa` (or other [test tokens](https://stripe.com/docs/testing#cards)) with `sk_test_...` keys,\nor omit `paymentMethod` and pass `savedCardId` from `POST /api/v1/cards` (same `customer.email` as when the card was saved).\nSet `STRIPE_PAYMENT_RETURN_URL` in `.env` if Stripe requires a return URL.\n",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            },
            "description": "Unique per logical payment; repeats within 24h return the same response. Swagger UI fills a new random UUID on each docs page load."
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePaymentRequest"
              },
              "examples": {
                "stripeTestTokVisa": {
                  "summary": "Stripe — test token (sk_test)",
                  "value": {
                    "provider": "stripe",
                    "amount": 2000,
                    "currency": "usd",
                    "description": "Swagger test payment",
                    "customer": {
                      "email": "test@example.com",
                      "name": "Test User"
                    },
                    "paymentMethod": {
                      "type": "card",
                      "token": "tok_visa"
                    }
                  }
                },
                "stripeSavedCard": {
                  "summary": "Stripe — saved card id from GET /cards",
                  "value": {
                    "provider": "stripe",
                    "amount": 2000,
                    "currency": "usd",
                    "description": "Pay with saved card",
                    "customer": {
                      "email": "test@example.com",
                      "name": "Test User"
                    },
                    "savedCardId": "REPLACE_WITH_ID_FROM_GET_CARDS"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Payment created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "502": {
            "$ref": "#/components/responses/ProviderError"
          }
        }
      },
      "get": {
        "tags": [
          "Payments"
        ],
        "summary": "List payments",
        "description": "Returns persisted **Mercado Pago** payments (from webhooks). Stripe/PayPal charges are not stored in this table yet.\nQuery: `status`, `externalReference`, `mpPaymentId`, `limit` (default 100, max 500), `offset`.\n",
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "example": "approved"
          },
          {
            "in": "query",
            "name": "externalReference",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "mpPaymentId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "in": "query",
            "name": "offset",
            "schema": {
              "type": "integer",
              "default": 0
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaymentListResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  }
}